+++ src/core/Popover/Wrapper.tsx
@@ -8,6 +8,7 @@ import {
 import { getCoordinates } from './../Popover/utils';
 import { useFadeTransition } from '../../components/composites/Transitions/useFadeTransition';
 import isEqual from 'lodash/isEqual';
+import { useKeyboardDismissable } from '../../hooks';
 
 type PopoverWrapperType = {
   popoverItem: any;
@@ -76,6 +77,16 @@ function Wrapper({
     }
   };
 
+  const handleClose = () => {
+    setPopoverItem(null);
+    popoverConfig.onClose && popoverConfig.onClose();
+  };
+
+  useKeyboardDismissable({
+    enabled: !!popoverItem && popoverConfig.isKeyboardDismissable,
+    onClose: handleClose,
+  });
+
   // DOC: Using this to get dimensions of trigger once it has been mounted.
   React.useEffect(() => {
     if (popoverConfig.triggerRef && popoverConfig.triggerRef.current) {
@@ -99,14 +110,10 @@ function Wrapper({
     <Animated.View
       style={[providerStyle.wrapper, { opacity: fadeValue }]}
       pointerEvents="box-none"
+      onAccessibilityEscape={handleClose}
     >
       {popoverItem && (
-        <TouchableWithoutFeedback
-          onPress={() => {
-            setPopoverItem(null);
-            popoverConfig.onClose && popoverConfig.onClose();
-          }}
-        >
+        <TouchableWithoutFeedback onPress={handleClose}>
           <View style={providerStyle.wrapper} />
         </TouchableWithoutFeedback>
       )}
