+++ src/hooks/index.ts
@@ -10,3 +10,4 @@ export { useToken } from './useToken';
 export { useSafeArea } from './useSafeArea';
 export { useContrastText } from './useContrastText';
 export { useScreenReaderEnabled } from './useScreenReaderEnabled';
+export { useKeyboardDismissable } from './useKeyboardDismissable';
