+++ src/hooks/useKeyboardDismissable.ts
@@ -1,0 +1,29 @@
+import React from 'react';
+import { Platform } from 'react-native';
+
+type IParams = {
+  enabled?: boolean;
+  onClose: () => void;
+};
+
+export const useKeyboardDismissable = ({ enabled, onClose }: IParams) => {
+  React.useEffect(
+    function closeOverlayOnEscapeEffectCallback() {
+      let escapeKeyListener: any = null;
+      if (Platform.OS === 'web') {
+        escapeKeyListener = (e: KeyboardEvent) => {
+          if (e.key === 'Escape' && enabled) {
+            onClose();
+          }
+        };
+        document.addEventListener('keydown', escapeKeyListener);
+      }
+      return () => {
+        if (Platform.OS === 'web') {
+          document.removeEventListener('keydown', escapeKeyListener);
+        }
+      };
+    },
+    [enabled, onClose]
+  );
+};
